+++ crux-mir/CHANGELOG.md
@@ -1,5 +1,9 @@
 # next
 
+This release supports [version
+9](https://github.com/GaloisInc/mir-json/blob/master/SCHEMA_CHANGELOG.md#9) of
+`mir-json`'s schema.
+
 # 0.12 -- 2026-01-29
 
 This release supports [version